requirements-macM1.txt
jupyterlab==3.2.9
metaflow
nbformat==5.4.0
notebook==6.4.10
numpy==1.23.1
pandas==1.3.5
matplotlib==3.5.2
scikit-learn==1.1.2
tensorflow-macos==2.7.0
xgboost==1.5.1
pyarrow==5.0.0
